Compiler warning cleanup: correct various complier warnings generated when additional warnings are turned on.  No significant logic changes
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@21291 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/readable.c
@@ -2073,7 +2073,7 @@ void tailor_readable_ob(object *book, int msg_type) {
     if (msg_type >= (int)arraysize(max_titles))
         msg_type = 0;
 
-    msg_type = msg_type > 0 ? msg_type : RANDOM()%6;
+    msg_type = msg_type > 0 ? msg_type : (int)(RANDOM()%6);
     switch (msg_type) {
     case MSGTYPE_MONSTER:
         message = mon_info_msg(level, book_buf_size, book);
